Temporarily remove release notes link and update file name to match others (case)
