+++ docs/infrastructure-as-code-management.md
@@ -1,14 +1,14 @@
 ---
 hide_table_of_contents: true
 hide_title: true
-title: Infrastrcuture as Code Management
+title: Infrastructure as Code Management
 # id: iacm
 ---
 
-<!-- # Infrastrcuture as Code Management -->
+<!-- # Infrastructure as Code Management -->
 
 <!-- Custom component -->
 
-import IaCM from '@site/src/components/Docs/iacm';
+import IaCM from '@site/src/components/Docs/Iacm';
 
 <IaCM />
